disk: Reality hits: delete on Close() in all cases of errors.

+++ caching/disk.go
@@ -392,6 +392,7 @@ type storageWriter struct {
 	path           string
 	invalidated    bool
 	errored        bool
+	deleted        bool
 	closeFinisher  func(name string, size int64)
 	closeNotifier  *chan Key
 	closed         bool
@@ -498,11 +499,13 @@ func (sw *storageWriter) Close() error {
 		if sw.wasRevalidated {
 			fd, err := os.OpenFile(sw.path, os.O_RDWR, 0)
 			if err != nil {
-				sw.log.Errorf("Could not reopen file for revalidation state saving: %v", err)
+				sw.log.Errorf("Could not reopen file %v for revalidation state saving: %v", sw.path, err)
+				sw.Delete()
 				return err
 			}
 			sm, err := getStorageMetadata(fd, metadataXAttrName)
 			if err != nil {
+				sw.Delete()
 				return err
 			}
 			sm.Revalidated = sw.now().Unix()
@@ -513,17 +516,17 @@ func (sw *storageWriter) Close() error {
 		}
 	}
 
-	err := sw.fd.Close()
+	fi, err := sw.fd.Stat()
 	if err != nil {
+		sw.log.Errorf("Could not stat file %v: %v", sw.fd.Name(), err)
+		sw.Delete()
 		return err
 	}
 	sizeOnDisk := fi.Size()
 
-	if sw.invalidated {
-		err := os.Remove(sw.path)
-		if err != nil {
-			sw.log.Warnf("Could not remove invalidated path %v: %v", sw.path, err)
-		}
+	err = sw.fd.Close()
+	if err != nil {
+		sw.Delete()
 		return err
 	}
 
@@ -580,11 +583,13 @@ func (sw *storageWriter) Close() error {
 
 	esm, err := encodeStorageMetadata(metadata)
 	if err != nil {
+		sw.Delete()
 		return err
 	}
 
 	err = xattr.Set(sw.path, metadataXAttrName, esm)
 	if err != nil {
+		sw.Delete()
 		return err
 	}
 
@@ -659,12 +664,18 @@ func (sw *storageWriter) SetRevalidated() {
 }
 
 func (sw *storageWriter) Delete() error {
+	if sw.deleted {
+		return nil
+	}
+
 	closeErr := sw.fd.Close()
 	err := os.Remove(sw.path)
 	if err != nil {
 		sw.log.Errorf("Could not remove path %v: %v. Close error was: %v", sw.path, err, closeErr)
 		return err
 	}
+
+	sw.deleted = true
 
 	return nil
 }
